Move renderButtons into its own file
